Implemented adding new players in the list for every player in lobby.
MenuController:
Removed the redundant storage of PlayerState, as it returns nullptr when a player joins the server.
Registered the MenuController class to the GameState's OnPlayerListUpdated delegate. This ensures the OnPlayerListUpdated method is triggered after a player joins, allowing the controller to update the player list widget dynamically.

GameInstance:
Added a getter for IsHost and implemented logic to set it to true when a player creates a session.
Ensures the lobby widget reference in the controller gets a valid value.
LobbyWidget:

Implemented the ClearPlayerList function to prevent duplicate players from being listed in the player list when new players join.

GameState:
Introduced a delegate to notify changes in the PlayerState array.
Created an array of MenuController to track players in the lobby.
Implemented OnRep_PlayerState() to replicate the updated player list to all clients and trigger notifications for player list changes.

MenuGameMode:
Added PostLogin and Logout logic to manage adding and removing players from the GameState's player list.

PropHuntPlayerController:
Changing input to game and show mouse crusor to false, when player game hud created.